backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
marshmallow==3.20.1
python-dotenv==1.0.0
pdfkit==1.0.0
